slight changes for image processing, further CSS work gor the groups page
git-svn-id: http://plugins.svn.wordpress.org/wp-e-commerce@118611 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ trunk/processing_functions.php
@@ -380,14 +380,14 @@ function wpsc_item_process_image($id, $input_file, $output_filename, $width = 0,
 
 
 				case 0:
-				$height = null;
-				$width  = null;
+				$height = (int)null;
+				$width  = (int)null;
 				break;
 
 				case 1:
 				default:
-				$height = get_option('product_image_height');
-				$width  = get_option('product_image_width');
+				$height = (int)get_option('product_image_height');
+				$width  = (int)get_option('product_image_width');
 				break;
 			}
 // 			if(($resize_method == 3) && ($_FILES['thumbnailImage'] != null) && file_exists($_FILES['thumbnailImage']['tmp_name'])) {
